+++ app/src/main/kotlin/ac/mdiq/podcini/utils/Logging.kt
@@ -3,11 +3,11 @@
 package ac.mdiq.podcini.utils
 
 import ac.mdiq.podcini.BuildConfig
+import ac.mdiq.podcini.net.download.DownloadError
 import ac.mdiq.podcini.storage.database.appPrefs
 import ac.mdiq.podcini.storage.database.runOnIOScope
 import ac.mdiq.podcini.storage.model.DownloadResult
 import ac.mdiq.podcini.storage.model.DownloadResult.Companion.logDownloadResult
-import ac.mdiq.podcini.storage.model.Episode
 import ac.mdiq.podcini.storage.model.Feed
 import android.util.Log
 import androidx.compose.runtime.getValue
@@ -48,12 +48,12 @@ fun Loge(t: String, m: String) {
     }
 }
 
-fun LogeFor(t: String, episode: Episode?, m: String) {
+fun LogeFor(t: String, episodeId: Long?, m: String) {
     if (BuildConfig.DEBUG || appPrefs.printDebugLogs) Log.e(t, m)
     LogScope.launch {
         trimToasts()
         if (appPrefs.showErrorToasts) toastMassege = "$t: Error: $m"
-        sessionLogs.add("${fullDateTimeString()} $t: ${episode?.id} Error: $m")
+        sessionLogs.add("${fullDateTimeString()} $t: $episodeId Error: $m")
     }
 }
 
@@ -67,13 +67,13 @@ fun Logs(t: String, m: String) {
     }
 }
 
-fun LogsFor(t: String, episode: Episode?, m: String) {
+fun LogsFor(t: String, episodeId: Long?, m: String) {
     if (BuildConfig.DEBUG || appPrefs.printDebugLogs) Log.e(t, m)
     showStackTrace()
     LogScope.launch {
         trimToasts()
         if (appPrefs.showErrorToasts) toastMassege = "$t: Error: $m "
-        sessionLogs.add("${fullDateTimeString()} $t: ${episode?.id} Error: $m ")
+        sessionLogs.add("${fullDateTimeString()} $t: $episodeId Error: $m ")
     }
 }
 
@@ -87,13 +87,13 @@ fun Logs(t: String, e: Throwable, m: String = "") {
     }
 }
 
-fun LogsFor(t: String, episode: Episode?, e: Throwable, m: String = "") {
+fun LogsFor(t: String, episodeId: Long?, e: Throwable, m: String = "") {
     if (BuildConfig.DEBUG || appPrefs.printDebugLogs) Log.e(t, m + ": "+ e.message + "\n" + Log.getStackTraceString(e))
     val me = e.message
     LogScope.launch {
         trimToasts()
         if (appPrefs.showErrorToasts) toastMassege = "$t: $m Error: $me"
-        sessionLogs.add("${fullDateTimeString()} $t: ${episode?.id} $m Error: $me")
+        sessionLogs.add("${fullDateTimeString()} $t: $episodeId $m Error: $me")
     }
 }
 
@@ -106,18 +106,19 @@ fun Logt(t: String, m: String) {
     }
 }
 
-fun LogtFor(t: String, episode: Episode?, m: String) {
+fun LogtFor(t: String, episodeId: Long?, m: String) {
     if (BuildConfig.DEBUG || appPrefs.printDebugLogs) Log.d(t, m)
     LogScope.launch {
         trimToasts()
         toastMassege = "$t: $m"
-        sessionLogs.add("${fullDateTimeString()} $t: ${episode?.id} $m")
+        sessionLogs.add("${fullDateTimeString()} $t: $episodeId $m")
     }
 }
 
-fun LogFor(t: String, feed: Feed, success: Boolean, message: String, toastAnyway: Boolean = false) {
-    runOnIOScope { logDownloadResult(DownloadResult(feed, null, success, message)) }
-    if (toastAnyway || !success) Logt(t, "Feed operation: success=$success, $message: ${feed.title}")
+fun LogFor(t: String, feed: Feed, success: Boolean, message: String, reason:  DownloadError? = null, toastAnyway: Boolean = false) {
+    runOnIOScope { logDownloadResult(DownloadResult(feed, reason, success, message)) }
+    if (toastAnyway && success) Logt(t, "Feed operation: success=$success, $message: ${feed.title}")
+    if (!success) Loge(t, "Feed operation: success=$success, $message: ${feed.title}")
 }
 
 fun showStackTrace() {